Manually toggle animation when switching layouts
Before, whether to animate or not was determined by an init variable.
If you were displaying a layout for the first time it would just show,
if not it would animate.

Now, you have to manually choose whether or not you want to animate.
This helps a lot when reacting to window resize events.